pyproject.toml
[project]
name = "seg_cityscapes"        # (optionnel) pour le nom du venv
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "juliengatineau", email = "julien_gatineau@hotmail.com" }]

dependencies = [
  "flask==3.0.3",
  "numpy==1.26.4",
  "matplotlib==3.8.4",
  "pillow==10.3.0",
  "requests==2.31.0",
  "torch==2.3.1+cpu",
  "torchvision==0.18.1+cpu",
  "huggingface-hub (>=0.34.4,<0.35.0)",
  "gunicorn (>=23.0.0,<24.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/torch_stable.html"
priority = "supplemental"   # <= important: utilisé en plus de PyPI

[tool.poetry]
package-mode = false

